% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formatting.R
\name{add_Formatting}
\alias{add_Formatting}
\title{Formatting the values.}
\usage{
add_Formatting(df)
}
\arguments{
\item{df}{A data frame produced by \code{SEASON_SUM_STATS3()} in stat ladders.}
}
\value{
The same data frame with diplay and output values.
}
\description{
Creates a display value and the output value.
}
