% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChampionStats.R
\name{load_Champion}
\alias{load_Champion}
\title{Load all Champion Data}
\usage{
load_Champion(
  pathChamp = system.file("extdata", "Champion_Stats_V.xlsx", package =
    "RichmondAnalytics")
)
}
\value{
A data frame: SSMET
}
\description{
Load all Champion Data
}
