% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricCats.R
\name{load_MetricCats}
\alias{load_MetricCats}
\title{Load all Champion Data}
\usage{
load_MetricCats(
  path = system.file("extdata", "Champion_Stats_V.xlsx", package = "RichmondAnalytics")
)
}
\value{
A list of data frames: CS, CS1, CS2, CS3
}
\description{
Load all Champion Data
}
