% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{load_Metrics}
\alias{load_Metrics}
\title{Load all Champion Data}
\usage{
load_Metrics(
  pathMetrics = system.file("extdata", "Metrics26.xlsx", package = "RichmondAnalytics")
)
}
\value{
A data frame: SSMET
}
\description{
Load all Champion Data
}
