% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeasonStats.R
\name{load_SeasonStats}
\alias{load_SeasonStats}
\title{Load all Champion Data}
\usage{
load_SeasonStats(
  path = system.file("extdata", "SEASON_SUM_STATS.xlsx", package = "RichmondAnalytics")
)
}
\value{
A data frame: SSMET
}
\description{
Load all Champion Data
}
